+++ src/NeatMapper.Async/IAsyncMapper.cs
@@ -1,4 +1,8 @@
-﻿namespace NeatMapper.Async {
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace NeatMapper.Async {
 	/// <summary>
 	/// Interface which allows asynchronously mapping an object to a new one or an existing one
 	/// </summary>
@@ -12,7 +16,7 @@
 		/// <param name="destinationType">type of the destination object to create, used to retrieve the available maps</param>
 		/// <param name="cancellationToken">cancellation token used to cancel async operations, will be forwarded to all the contexts in the mapping</param>
 		/// <returns>a task which when completed returns the newly created object of <paramref name="destinationType"/>, which may be null</returns>
-		public Task<
+		Task<
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
 			object?
 #else
@@ -44,7 +48,7 @@
 		/// a task which when completed returns the resulting object of the mapping of <paramref name="destinationType"/> type,
 		/// which can be the same as <paramref name="destination"/> or a new one, may be null
 		/// </returns>
-		public Task<
+		Task<
 #if NETCOREAPP3_1_OR_GREATER || NETSTANDARD2_1_OR_GREATER || NET5_0_OR_GREATER
 			object?
 #else