+++ src/NeatMapper.Async/Internal/TaskUtils.cs
@@ -1,9 +1,13 @@
-﻿namespace NeatMapper.Async.Internal {
+﻿using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace NeatMapper.Async.Internal {
 	/// <summary>
 	/// Helper class to await Tasks and return the result vie reflection
 	/// </summary>
 	internal static class TaskUtils {
-		public static async Task<TResult?> AwaitTask<TResult>(Task task) {
+		public static async Task<TResult> AwaitTask<TResult>(Task task) {
 			await task;
 			if (task.IsFaulted) {
 				if(task.Exception != null){
@@ -19,7 +23,7 @@
 				throw new TaskCanceledException();
 			else {
 				try {
-					return (TResult)task.GetType().GetProperty(nameof(Task<object>.Result))!.GetValue(task)!;
+					return (TResult)task.GetType().GetProperty(nameof(Task<object>.Result)).GetValue(task);
 				}
 				catch {
 					throw new Exception("Task contains no result");