+++ src/NeatMapper.Common/CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
-## [1.2.0] - Unreleased
+## [2.0.0] - Unreleased
 
 ### Added
 
-- New interface `IHierarchyMatchMap<TSource, TDestination>` (and its .NET 7+ static counterpart `IHierarchyMatchMapStatic<TSource, TDestination>`) which allows matching two types as well as derived types, will be automatically used when merging collections.
+- `IHierarchyMatchMap<TSource, TDestination>` (and its .NET 7+ static counterpart `IHierarchyMatchMapStatic<TSource, TDestination>`) which allows matching two types as well as derived types, will be automatically used when merging collections.
+
+### Removed
+
+- `IMapperConfiguration`
 
 ## [1.1.0] - 2023-10-02