+++ src/NeatMapper.DependencyInjection/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.0] - Unreleased
+
+### Changed
+
+- NeatMapper dependency version
+
 ## [1.0.2] - 2023-10-02
 
 ### Fixed